Fiy Next adapter build on windows
