+++ inlang/source-code/paraglide/paraglide-js-adapter-next/build.mjs
@@ -1,10 +1,5 @@
 import { rollup } from "rollup"
-import preserveDirectives from "rollup-preserve-directives"
-import typescript from "@rollup/plugin-typescript"
-import cjs from "@rollup/plugin-commonjs"
-import json from "@rollup/plugin-json"
-import resolve from "@rollup/plugin-node-resolve"
-import replace from "@rollup/plugin-replace"
+import { plugins } from "./plugins.mjs"
 import fs from "node:fs/promises"
 
 //remove the old dist folder
@@ -18,93 +13,76 @@ if (typeof version !== "string") {
 const peerDependencies = Object.keys(packageJson.peerDependencies || {})
 const dependencies = Object.keys(packageJson.dependencies || {})
 
-export const plugins = [
-	typescript({ tsconfig: "./tsconfig.json" }),
-	cjs(),
-	resolve(),
-	preserveDirectives(),
-	json(),
-	replace({
-		values: {
-			MARKETPLACE_ID: JSON.stringify("library.inlang.paraglideJsAdapterNextJs"),
-			PARAGLIDE_NEXT_VERSION: JSON.stringify(version),
-		},
-		preventAssignment: true,
-	}),
+// eslint-disable-next-line no-undef
+// eslint-disable-next-line no-console
+// eslint-disable-next-line no-undef
+console.info(`Building ${packageJson.name} v${packageJson.version}...`)
+
+const external = [
+	/node_modules/,
+	/^node:/,
+	"$paraglide/runtime.js",
+	...peerDependencies,
+	...dependencies,
 ]
 
-// eslint-disable-next-line no-undef
-if (process.env.SCRIPT === "build") {
-	// eslint-disable-next-line no-console
-	// eslint-disable-next-line no-undef
-	console.info(`Building ${packageJson.name} v${packageJson.version}...`)
+const app_build = await rollup({
+	plugins,
+	input: {
+		"app/index.server": "src/app/index.server.tsx",
+		"app/index.client": "src/app/index.client.tsx",
+	},
+	external,
+})
 
-	const external = [
-		/node_modules/,
-		/^node:/,
-		"$paraglide/runtime.js",
-		...peerDependencies,
-		...dependencies,
-	]
+const pages_build = await rollup({
+	plugins,
+	input: {
+		"pages/index": "src/pages/index.tsx",
+	},
+	external,
+})
 
-	const app_build = await rollup({
-		plugins,
-		input: {
-			"app/index.server": "src/app/index.server.tsx",
-			"app/index.client": "src/app/index.client.tsx",
-		},
-		external,
-	})
+const pluginBuild = await rollup({
+	plugins,
+	input: {
+		"plugin/index": "src/plugin/index.ts",
+	},
+	external,
+})
 
-	const pages_build = await rollup({
-		plugins,
-		input: {
-			"pages/index": "src/pages/index.tsx",
-		},
-		external,
-	})
+const cliBuild = await rollup({
+	plugins,
+	input: {
+		"cli/index": "src/cli/index.ts",
+	},
+	external: [/^node:/, ...peerDependencies, ...dependencies],
+})
 
-	const pluginBuild = await rollup({
-		plugins,
-		input: {
-			"plugin/index": "src/plugin/index.ts",
-		},
-		external,
-	})
+await app_build.write({
+	preserveModules: true,
+	format: "es",
+	entryFileNames: "[name].js",
+	dir: "dist",
+})
 
-	const cliBuild = await rollup({
-		plugins,
-		input: {
-			"cli/index": "src/cli/index.ts",
-		},
-		external: [/^node:/, ...peerDependencies, ...dependencies],
-	})
+await pages_build.write({
+	preserveModules: false,
+	format: "cjs",
+	entryFileNames: "[name].js",
+	dir: "dist",
+})
 
-	await app_build.write({
-		preserveModules: true,
-		format: "es",
-		entryFileNames: "[name].js",
-		dir: "dist",
-	})
+await pluginBuild.write({
+	preserveModules: false,
+	format: "cjs",
+	entryFileNames: "[name].cjs",
+	dir: "dist",
+})
 
-	await pages_build.write({
-		preserveModules: false,
-		format: "cjs",
-		entryFileNames: "[name].js",
-		dir: "dist",
-	})
-
-	await pluginBuild.write({
-		preserveModules: false,
-		format: "cjs",
-		entryFileNames: "[name].cjs",
-		dir: "dist",
-	})
-
-	await cliBuild.write({
-		preserveModules: true,
-		format: "esm",
-		entryFileNames: "[name].mjs",
-		dir: "dist",
-	})
-}
+await cliBuild.write({
+	preserveModules: true,
+	format: "esm",
+	entryFileNames: "[name].mjs",
+	dir: "dist",
+})
